+++ src/lib.rs
@@ -22,24 +22,14 @@ pub mod ssh;
 pub mod cloud_exec;
 #[cfg(feature = "k8s-exec")]
 pub mod k8s_exec;
-#[cfg(feature = "mqtt")]
-pub mod mqtt_exec;
-#[cfg(feature = "nats")]
-pub mod nats_exec;
-#[cfg(feature = "netconf")]
-pub mod netconf;
 #[cfg(feature = "serial")]
 pub mod serial_port;
-#[cfg(feature = "snmp")]
-pub mod snmp_client;
 #[cfg(feature = "ssm")]
 pub mod ssm;
 #[cfg(feature = "telnet")]
 pub mod telnet;
 #[cfg(feature = "winrm")]
 pub mod winrm;
-#[cfg(feature = "zeromq")]
-pub mod zmq_exec;
 
 pub use config::Config;
 pub use error::{BridgeError, Result};